Helperを作成し、ImagePickerを別ファイルに分けた
