+++ ImageSaveSampleApp/Helper/ImagePicker.swift
@@ -1,0 +1,40 @@
+//
+//  ImagePicker.swift
+//  ImageSaveSampleApp
+//
+//  Created by RikutoSato on 2023/12/19.
+//
+
+import SwiftUI
+
+struct ImagePicker: UIViewControllerRepresentable {
+    @Binding var imageData: Data?
+    
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        let picker = UIImagePickerController()
+        picker.delegate = context.coordinator
+        return picker
+    }
+    
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+    
+    class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
+        let parent: ImagePicker
+        
+        init(_ parent: ImagePicker) {
+            self.parent = parent
+        }
+        
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+            if let image = info[.originalImage] as? UIImage {
+                parent.imageData = image.jpegData(compressionQuality: 0.5)
+            }
+            
+            picker.dismiss(animated: true)
+        }
+    }
+    
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
+}
